feat(Application): Add types to export
