+++ src/Application/Application.ts
@@ -10,20 +10,7 @@ import {
   ContainerOptions,
 } from 'awilix';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type AsyncResolver<T> = (cradle: any) => Promise<Resolver<T>>;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export interface Provider<T = any> {
-  name: string;
-  resolver: Resolver<T>;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export interface AsyncProvider<T = any> {
-  name: string;
-  asyncResolver: AsyncResolver<T>;
-}
+import { AsyncProvider, AsyncResolver, Provider } from './types';
 
 const isAsyncProvider = (provider: Provider | AsyncProvider): provider is AsyncProvider =>
   !!(provider as AsyncProvider).asyncResolver;
