+++ src/Application/index.ts
@@ -1,3 +1,6 @@
 import Application from './Application';
+import { AsyncProvider, AsyncResolver, Provider } from './types';
 
 export default Application;
+
+export { AsyncProvider, AsyncResolver, Provider };
